+++ shaders/program/gbuffers_clouds.glsl
@@ -2,6 +2,9 @@
 	in_out vec2 texcoord;
 #endif
 in_out vec3 playerPos;
+#if STORY_MODE_CLOUDS_ENABLED == 1
+	in_out float alphaMult;
+#endif
 
 
 
@@ -33,6 +36,22 @@ void main() {
 	// nearby transparency
 	float playerPosDist = max(length(playerPos.xz), abs(playerPos.y));
 	color.a = 1.0 - mix(NEARBY_CLOUD_TRANSPARENCY, VANILLA_CLOUD_TRANSPARENCY, clamp(playerPosDist / NEARBY_CLOUD_DIST, 0.0, 1.0));
+	
+	
+	// story mode clouds
+	#if STORY_MODE_CLOUDS_ENABLED == 1
+		float alphaMult = 1.0 - alphaMult;
+		#if STORY_MODE_CLOUDS_CURVE == 2
+			alphaMult = pow2(alphaMult);
+		#elif STORY_MODE_CLOUDS_CURVE == 3
+			alphaMult = pow3(alphaMult);
+		#elif STORY_MODE_CLOUDS_CURVE == 4
+			alphaMult = pow4(alphaMult);
+		#elif STORY_MODE_CLOUDS_CURVE == 5
+			alphaMult = pow5(alphaMult);
+		#endif
+		color.a *= 1.0 - alphaMult;
+	#endif
 	
 	
 	// cloud color
@@ -93,7 +112,12 @@ void main() {
 	#endif
 	
 	vec3 viewPos = transform(gl_ModelViewMatrix, gl_Vertex.xyz);
-	playerPos = mat3(gbufferModelViewInverse) * viewPos;
+	playerPos = transform(gbufferModelViewInverse, viewPos);
+	#if STORY_MODE_CLOUDS_ENABLED == 1
+		alphaMult = percentThrough(cameraPosition.y, cloudHeight + 32.0, cloudHeight - 8.0); // start off inverted
+		alphaMult *= float(playerPos.y + cameraPosition.y > cloudHeight + 1);
+		alphaMult = 1.0 - alphaMult;
+	#endif
 	
 	gl_Position = viewToNdc(viewPos);